acs: Updates for the optimized quote flow
The newer IMA specification guarantees that the append/extend
operation is atomic.  This permits attestation flow optimization.  The
quoted PCRs can always be reconstructed from the logs.

The logs now come with the quote.  PCRs are not sent, and the expected
PCRs are derived from the logs.  This eliminates the pcrread loop.

Signed-off-by: Ken Goldman <kgoldman@us.ibm.com>